LABS/Airflow_LAB/weatherflow/requirements.txt
# Core Dependencies
apache-airflow==2.10.3
apache-airflow-providers-postgres==5.11.0
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0

# Database
psycopg2-binary
sqlalchemy

# Data Processing
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.11.0
pylint==3.0.2
flake8==6.1.0

# Phase 2: ML Dependencies (uncomment when needed)
# tensorflow==2.15.0
# scikit-learn==1.3.2
# prophet==1.1.5
# mlflow==2.8.1

# Phase 3: Dashboard Dependencies (uncomment when needed)
# streamlit==1.28.2
# plotly==5.18.0
# folium==0.15.0
